This version uses floats instead of truncating integers

+++ 1_2.c
@@ -3,14 +3,13 @@
 int main(){
 
     //temp refers to temperature
-    int f_temp, c_temp;
+    float f_temp, c_temp;
 
-    //print a table of fahrenheit and celsius equivalents. All operations are rounded to the nearest integer.
+    //print a table of fahrenheit and celsius equivalents.
     f_temp = 0;
     while (f_temp <= 300) {
         c_temp = 5 * (f_temp - 32) / 9;;
-        printf("%3d\t%3d\n", f_temp, c_temp);
-        //adding the 3 after the % makes the output right justified
+        printf("%3.1f\t%3.1f\n", f_temp, c_temp);
         f_temp += 20;
     }
